Update patch set 1
Patch Set 1:

wiping and flashing in this recovery works, but there are issues with the output recovery that we might as well fix right now.

issue 1: the option to mount sdcard doesn't show up in the menu
issue 2: display only updates after every 2nd volume click.

i'll be investigating, and will submit a new patchset when i find the issue.

Patch-set: 1