Update patch set 1
Patch Set 1:

Ameer, wow, you have nerves of steel.  I wasn't expecting anyone to actually try this on a device.

So when you formatted and installed a rom using it, had you also applied this patch?  http://review.cyanogenmod.com/17326

If you did apply that patch, I will wait to merge this in until 17326 is approved.  If you skipped that patch, then the GB kernel was enough to stop the corruption and I will go ahead and merge this patch in.

Patch-set: 1