Update patch set 1
Patch Set 1: Verified; Looks good to me, approved

Going ahead and merging this in to put as much preventative measures in place as possible. We can address the remaining recovery bugs in separate patches.

Obviously, when we get kernel source the GB kernel in this patch will go away.

Patch-set: 1